implemented huge

+++ include/strategies/slab.h
@@ -4,10 +4,10 @@ typedef struct
 {
     char stratigy_id;
     void* heap_start;
-    size_t object_size;
+    uint32_t object_size;
     void* free_list;
 } SlabAllocator;
-void slab_init(SlabAllocator *s, size_t object_size);
+void slab_init(SlabAllocator *s, uint32_t object_size);
 void* slab_alloc(SlabAllocator *s);
 void slab_free(SlabAllocator *s);
 void slab_destroy(SlabAllocator *s);
